Integration of hostapd with AP scripts

+++ ap/ap-off.sh
@@ -5,25 +5,32 @@
 # Kill off the monomanaical DNS server
 #
 echo ":: Shutting down fake DNS Server"
-kill -9 `cat /run/monodns.pid`
-rm /run/monodns.pid
+if [ -e /run/monodns.pid ]
+then
+    kill -9 `cat /run/monodns.pid`
+    rm /run/monodns.pid
+fi
 #
-echo ":: Shutting down DHCP Server"
-killall dhcpd
-
+# Kill of DHCP Server if running
+echo ":: DHCP Server shutting down"
+if [ -e /run/dhcpd.pid ]
+then
+    kill `cat /run/dhcpd.pid`
+    rm /run/dhcpd.pid
+fi
+#
+# Kill hostapd
+echo ":: Hostapd Server shutting down"
+if [ -e /run/hostapd.pid ]
+then
+        kill `cat /run/hostapd.pid`
+        # rm /run/hostapd.pid # apparently this happens automagically
+fi
 #
 # Kill off wpa_supplicant
 #
 echo ":: WPA_Supplicant shutting down"
 killall wpa_supplicant
-#
-# Kill of DHCP Server if running
-echo ":: DHCP Server shutting down"
-killall dhcpd
-#
-# Kill hostapd
-echo ":: Hostapd Server shutting down"
-killall hostapd
 #
 # Turn the link off
 #
@@ -34,11 +41,10 @@ sleep 1
 # Reconfigure the link
 #
 #echo ":: Reconfiguring wlan0"
-#iwconfig wlan0 mode Managed
-#ip addr del 192.168.23.254/24 dev wlan0
-#ip route del default via 192.168.23.254
+#
+ip addr del 192.168.23.254/24 dev wlan0
+ip route del default via 192.168.23.254
 #
 ip link set up wlan0
 sleep 1
 echo ":: Access Point stopped"
-
